refactor(shared): extract segment focus navigation shared by date/time fields
The currentSegmentIndex / nextFocusableSegment / prevFocusableSegment /
focusNext block was duplicated across DateFieldRoot, DateRangeFieldRoot,
TimeFieldRoot and TimeRangeFieldRoot, which is how the RTL auto-advance
bug fixed in #2813 came to exist in four places. Extract it into a
useSegmentNavigation composable in shared/date/, parameterized by the
data attributes that identify a segment (range roots pass their extra
segment-type attribute to keep start/end disambiguation). No behavior
change intended: the logic is moved verbatim.
